[Parse] Move SyntaxParsingContext to Parse
This is only for the Parser.
Also, this resolves layering violation where libAST and libSyntax
depends on each other.
